third_party/blink/web_tests/animations/stability/unstringable-keyframe-value-crash.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_throws_js(TypeError, () => {
    document.body.animate({left: [{toString:null}]});
  });
}, 'Do not crash when passing an unstringable value in a property indexed keyframe');
</script>
